refactor(tests): split CommandVisibility.test.ts into 9 domain-focused files
Split the largest test file (2588 LOC) into smaller, maintainable files under 500 LOC each:

- CommandVisibility.conversion.test.ts (104 LOC) - task/project conversion
- CommandVisibility.creation.test.ts (353 LOC) - task/project/area/event creation
- CommandVisibility.effortPlanning.test.ts (318 LOC) - planning and trash
- CommandVisibility.instance.test.ts (200 LOC) - prototype instantiation
- CommandVisibility.maintenance.test.ts (439 LOC) - archive/clean/repair
- CommandVisibility.status.test.ts (356 LOC) - status transitions
- CommandVisibility.statusRollback.test.ts (162 LOC) - status rollback
- CommandVisibility.taskCreation.test.ts (413 LOC) - related tasks/daily notes
- CommandVisibility.voting.test.ts (270 LOC) - effort voting

All 200 tests pass. Total: 2615 LOC across 9 files (avg 290 LOC/file).

Partial progress on issue #474. Remaining large files:
- DailyTasksRenderer.test.ts (2560 LOC)
- CommandManager.test.ts (1913 LOC)
- ButtonGroupsBuilder.test.ts (1365 LOC)
- TaskCreationService.test.ts (1299 LOC)
- RelationsRenderer.test.ts (1148 LOC)
- ObsidianVaultAdapter.test.ts (1022 LOC)
